requirements.txt
-i https://pypi.org/simple
asgiref==3.6.0 ; python_version >= '3.7'
certifi==2022.12.7 ; python_version >= '3.6'
charset-normalizer==2.1.1 ; python_full_version >= '3.6.0'
django==4.1.5
django-cors-headers==3.13.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
idna==3.4 ; python_version >= '3.5'
pillow==9.4.0
pyjwt==2.6.0 ; python_version >= '3.7'
pytz==2022.7
requests==2.28.1
sqlparse==0.4.3 ; python_version >= '3.5'
tzdata==2022.7 ; sys_platform == 'win32'
urllib3==1.26.14 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
